move runQueryViaBatch to utilities
